Add files via upload
This project encompasses data pertaining to the Iris plant species. Utilising the provided characteristics, specifically measurements, for each species, the primary objective is to construct a classification model. The aim of this model is to facilitate the identification of Iris plant species based solely on their measurements, employing machine learning methodologies.

>I have used Logistic Regression model